Wallet: transactios shimmers and some cleaning
