--- lib/v2/screens/wallet/components/transactions/components/transaction_loadign_card.dart
@@ -1,16 +1,0 @@
-import 'package:flutter/material.dart';
-
-class TransactionLoadingCard extends StatelessWidget {
-  const TransactionLoadingCard({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.only(top: 10, bottom: 10, right: 20, left: 20),
-      child: Row(
-        children: [],
-      ),
-    );
-  }
-}
-//360 x 80
