--- lib/v2/screens/wallet/components/transactions/transactions_list_widget.dart
@@ -1,80 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:seeds/v2/constants/app_colors.dart';
-import 'package:seeds/v2/datasource/local/settings_storage.dart';
-import 'package:seeds/v2/domain-shared/page_state.dart';
-import 'package:seeds/v2/screens/wallet/components/transactions/components/transaction_info_card.dart';
-import 'package:seeds/v2/screens/wallet/components/transactions/interactor/viewmodels/transactions_list_bloc.dart';
-import 'package:seeds/v2/screens/wallet/components/transactions/interactor/viewmodels/transactions_list_events.dart';
-import 'package:seeds/v2/screens/wallet/components/transactions/interactor/viewmodels/transactions_list_state.dart';
-import 'package:seeds/v2/screens/wallet/interactor/viewmodels/bloc.dart';
-import 'package:shimmer/shimmer.dart';
-
-class TransactionsListWidget extends StatefulWidget {
-  @override
-  _TransactionsListWidgetState createState() => _TransactionsListWidgetState();
-}
-
-class _TransactionsListWidgetState extends State<TransactionsListWidget> with AutomaticKeepAliveClientMixin {
-  @override
-  Widget build(BuildContext context) {
-    super.build(context);
-    return BlocProvider<TransactionsListBloc>(
-      create: (_) => TransactionsListBloc()..add(OnLoadTransactionsList()),
-      child: BlocListener<WalletBloc, WalletState>(
-        listenWhen: (context, state) => state.pageState == PageState.loading,
-        listener: (context, state) {
-          BlocProvider.of<TransactionsListBloc>(context).add(OnLoadTransactionsList());
-        },
-        child: BlocBuilder<TransactionsListBloc, TransactionsListState>(
-          builder: (context, state) {
-            if (state.isLoadingNoData) {
-              return Column(
-                children: [for (var i = 0; i < 5; i++) loadingShimmer()],
-              );
-            } else {
-              return ListView.builder(
-                physics: const NeverScrollableScrollPhysics(),
-                shrinkWrap: true,
-                itemBuilder: (ctx, index) {
-                  var model = state.transactions[index];
-                  return TransactionInfoCard(
-                    callback: () {
-                      // TODO(n13): Implement callback - show tx detail
-                      print("Not implemented");
-                    },
-                    profileAccount: model.to == settingsStorage.accountName ? model.from : model.to,
-                    timestamp: model.timestamp,
-                    amount: model.quantity,
-                    incoming: settingsStorage.accountName == model.to,
-                  );
-                },
-                itemCount: state.transactions.length,
-              );
-            }
-          },
-        ),
-      ),
-    );
-  }
-
-  Widget loadingShimmer() => Shimmer.fromColors(
-        baseColor: AppColors.grey1,
-        highlightColor: AppColors.grey2,
-        child: Row(
-          children: <Widget>[
-            Expanded(
-              child: Container(
-                height: 44,
-                color: AppColors.black50,
-                margin: const EdgeInsets.only(left: 16, right: 16, bottom: 10),
-              ),
-            ),
-          ],
-        ),
-      );
-
-  @override
-  bool get wantKeepAlive => true;
-}
