+++ lib/v2/screens/wallet/components/transactions_list/components/transaction_loadign_card.dart
@@ -1,0 +1,43 @@
+import 'package:flutter/material.dart';
+import 'package:seeds/v2/components/shimmer_circle.dart';
+import 'package:seeds/v2/components/shimmer_rectangle.dart';
+
+class TransactionLoadingCard extends StatelessWidget {
+  const TransactionLoadingCard({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.only(top: 10, bottom: 10, right: 20, left: 20),
+      child: Row(
+        children: [
+          const ShimmerCircle(60),
+          Expanded(
+            child: Padding(
+              padding: const EdgeInsets.only(left: 10),
+              child: Column(
+                children: [
+                  Row(
+                    children: [
+                      const ShimmerRectangle(size: Size(110, 21)),
+                      const Spacer(),
+                      const ShimmerRectangle(size: Size(64, 21)),
+                    ],
+                  ),
+                  const SizedBox(height: 10),
+                  Row(
+                    children: [
+                      const ShimmerRectangle(size: Size(70, 16)),
+                      const Spacer(),
+                      const ShimmerRectangle(size: Size(44, 16)),
+                    ],
+                  )
+                ],
+              ),
+            ),
+          )
+        ],
+      ),
+    );
+  }
+}
