+++ lib/v2/screens/wallet/components/transactions_list/interactor/mappers/member_state_mapper.dart
@@ -1,6 +1,6 @@
 import 'package:async/async.dart';
 import 'package:seeds/v2/domain-shared/page_state.dart';
-import 'package:seeds/v2/screens/wallet/components/transactions/interactor/viewmodels/member_state.dart';
+import 'package:seeds/v2/screens/wallet/components/transactions_list/interactor/viewmodels/member_state.dart';
 
 class MemberStateMapper {
   MemberState mapResultToState(MemberState currentState, Result result) {
