+++ lib/v2/screens/wallet/components/transactions_list/interactor/mappers/transactions_state_mapper.dart
@@ -1,6 +1,6 @@
 import 'package:seeds/v2/domain-shared/page_state.dart';
 import 'package:async/async.dart';
-import 'package:seeds/v2/screens/wallet/components/transactions/interactor/viewmodels/transactions_list_state.dart';
+import 'package:seeds/v2/screens/wallet/components/transactions_list/interactor/viewmodels/transactions_list_state.dart';
 
 class TransactionsListStateMapper {
   TransactionsListState mapResultToState(TransactionsListState currentState, Result transactionsResult) {
