+++ lib/v2/screens/wallet/components/transactions_list/interactor/viewmodels/member_state.dart
