+++ lib/v2/screens/wallet/components/transactions_list/interactor/viewmodels/transactions_list_bloc.dart
@@ -1,9 +1,9 @@
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:seeds/v2/domain-shared/page_state.dart';
-import 'package:seeds/v2/screens/wallet/components/transactions/interactor/mappers/transactions_state_mapper.dart';
-import 'package:seeds/v2/screens/wallet/components/transactions/interactor/usecases/load_transactions_use_case.dart';
-import 'package:seeds/v2/screens/wallet/components/transactions/interactor/viewmodels/transactions_list_events.dart';
-import 'package:seeds/v2/screens/wallet/components/transactions/interactor/viewmodels/transactions_list_state.dart';
+import 'package:seeds/v2/screens/wallet/components/transactions_list/interactor/mappers/transactions_state_mapper.dart';
+import 'package:seeds/v2/screens/wallet/components/transactions_list/interactor/usecases/load_transactions_use_case.dart';
+import 'package:seeds/v2/screens/wallet/components/transactions_list/interactor/viewmodels/transactions_list_events.dart';
+import 'package:seeds/v2/screens/wallet/components/transactions_list/interactor/viewmodels/transactions_list_state.dart';
 
 class TransactionsListBloc extends Bloc<TransactionsListEvent, TransactionsListState> {
   TransactionsListBloc() : super(TransactionsListState.initial());
